% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_operator}
\alias{f_operator}
\title{Function operator for sqrt.2 and log.2}
\usage{
f_operator(f)
}
\arguments{
\item{f}{sqrt.2/log.2}
}
\value{
sqrt(x)/log(x) or error condition object,
with "invalid_input" subclass and invalid value attached
}
\description{
Function operator for sqrt.2 and log.2
}
\examples{
f_operator(sqrt.2)(2) = sqrt(2)
}
