% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log.2}
\alias{log.2}
\title{log.2}
\usage{
\method{log}{`2`}(x)
}
\arguments{
\item{x}{numeric value which you want to take log of}
}
\value{
log(x) or error with negative input
}
\description{
log.2
}
\examples{
log.2(2) = log(2)
}
