% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sqrt.2}
\alias{sqrt.2}
\title{Sqrt.2}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{numeric value which you want to take square root of}
}
\value{
sqrt(x) or error with negative input
}
\description{
Sqrt.2
}
\examples{
sqrt.2(2) = sqrt(2)
}
